Move private fs structures into fs_internal.h
